Reorganize project to be logical
