+++ Microsoft.EntityFrameworkCore.UnitOfWork.Host/Properties/launchSettings.json
