+++ Microsoft.EntityFrameworkCore.UnitOfWork/Collections/IPagedList.cs
@@ -2,7 +2,7 @@
 
 using System.Collections.Generic;
 
-namespace Microsoft.EntityFrameworkCore
+namespace Microsoft.EntityFrameworkCore.UnitOfWork.Collections
 {
     /// <summary>
     /// Provides the interface(s) for paged list of any type.
