+++ Microsoft.EntityFrameworkCore.UnitOfWork/Collections/IQueryablePageListExtensions.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.EntityFrameworkCore
+namespace Microsoft.EntityFrameworkCore.UnitOfWork.Collections
 {
     public static class IQueryablePageListExtensions
     {
