+++ Microsoft.EntityFrameworkCore.UnitOfWork/IRepository.cs
@@ -7,6 +7,7 @@ using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore.Query;
+using Microsoft.EntityFrameworkCore.UnitOfWork.Collections;
 
 namespace Microsoft.EntityFrameworkCore
 {
