+++ Microsoft.EntityFrameworkCore.UnitOfWork/Repository.cs
@@ -9,6 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Query;
+using Microsoft.EntityFrameworkCore.UnitOfWork.Collections;
 
 namespace Microsoft.EntityFrameworkCore
 {
